Finish the logic of the Tic Tac Toe game
There are 3 objects to control respective part of
the game:
- Gameboard controls the gameboard of the game
which is a board of 'x' and 'o'.
- Player is responsible for producing player data
including their names and markers.
- GameController controls the flow of the game,
it helps add the player data to the game flow and
play the game in turn. It is also able to check if
one of the players wins or the game is tie.

Now the game is able to run in console