Combine the logic with the UI
Now the player can tab on one of the 9 spots to
play. By default, the player with 'x' mark starts
first. After the game is finished (tie or not),
a dialog modal show up and inform the result with
a button to start a new game. They can also start
a new game by clicking the reset button below the
game board.